% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{ar_alla_lan_i_sverige}
\alias{ar_alla_lan_i_sverige}
\title{Innehåller regionkoderna alla 21 län i Sverige?}
\usage{
ar_alla_lan_i_sverige(
  reg_koder,
  tillat_rikskod = TRUE,
  returnera_text = FALSE,
  returtext = NA
)
}
\arguments{
\item{reg_koder}{Regionkoderna som testas.}

\item{tillat_rikskod}{Tillåt att rikskoden \code{"00"} finns med.}

\item{returnera_text}{Returnera \code{"Sveriges län"} i stället för \code{TRUE}.}

\item{returtext}{Text att returnera i stället för \code{FALSE} när villkoret inte
är uppfyllt och \code{returnera_text = TRUE}.}
}
\value{
\code{TRUE}/\code{FALSE}, eller en textsträng om \code{returnera_text = TRUE}.
}
\description{
Innehåller regionkoderna alla 21 län i Sverige?
}
