% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verktyg.R
\name{avrundning_dynamisk}
\alias{avrundning_dynamisk}
\title{Dynamisk avrundning}
\usage{
avrundning_dynamisk(
  x,
  grans_stora = 10,
  grans_medel = 1,
  dec_stora = 0,
  dec_medel = 1,
  dec_sma = 2
)
}
\arguments{
\item{x}{Numerisk vektor.}

\item{grans_stora, grans_medel}{Gränser för stora respektive medelstora tal.}

\item{dec_stora, dec_medel, dec_sma}{Antal decimaler i respektive intervall.}
}
\value{
En numerisk vektor.
}
\description{
Stora tal rundas till en "fin" grund (10, 100, ...), medelstora till en
decimal och små till två.
}
