% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{desokoder_bearbeta}
\alias{desokoder_bearbeta}
\title{Utöka läns-/kommunkoder till DeSO-koder}
\usage{
desokoder_bearbeta(
  tabell_id,
  koder,
  region_var = "Region",
  behall_bara_desokoder = TRUE,
  base_url = "https://statistikdatabasen.scb.se/api/v2/tables/"
)
}
\arguments{
\item{tabell_id}{PxWeb-tabell-id.}

\item{koder}{Läns- (2 tecken), kommun- (4 tecken) eller tätortskoder
(>4 tecken). \code{"*"} returneras oförändrat.}

\item{region_var}{Regionvariabelns namn i tabellen.}

\item{behall_bara_desokoder}{Om \code{TRUE} behålls bara koder längre än 4 tecken.}

\item{base_url}{Bas-URL till PxWeb API v2.}
}
\value{
En teckenvektor med DeSO-koder.
}
\description{
Utöka läns-/kommunkoder till DeSO-koder
}
