% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{excelfil_spara_formaterad}
\alias{excelfil_spara_formaterad}
\title{Spara en eller flera data.frames som formaterad xlsx}
\usage{
excelfil_spara_formaterad(
  indata,
  output_mapp = utskriftsmapp(),
  excelfil_namn = NA,
  auto_kolumnbredd = TRUE,
  fetstil_rader = 1,
  fetstil_kolumner = NA,
  filnamnstillagg = "xlsx",
  skriv_over_fil = TRUE
)
}
\arguments{
\item{indata}{En \code{data.frame} eller en (namngiven) lista av \code{data.frame}s.}

\item{output_mapp}{Målmapp.}

\item{excelfil_namn}{Filnamn. \code{NA} = härled från \code{indata}.}

\item{auto_kolumnbredd}{Anpassa kolumnbredd efter innehåll.}

\item{fetstil_rader, fetstil_kolumner}{Rader/kolumner i fetstil, eller \code{NA}.}

\item{filnamnstillagg}{Filändelse om \code{excelfil_namn} härleds.}

\item{skriv_over_fil}{Skriv över befintlig fil.}
}
\value{
Inget - skriver filen.
}
\description{
Spara en eller flera data.frames som formaterad xlsx
}
