% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webb.R
\name{filhamtning_med_url_och_sokord}
\alias{filhamtning_med_url_och_sokord}
\title{Hitta och ladda ner en fil från en webbsida via sökord}
\usage{
filhamtning_med_url_och_sokord(
  url_webbsida =
    "https://arbetsformedlingen.se/statistik/sok-statistik/tidigare-statistik",
  sokord = c("web-platser", ".xlsx"),
  bas_url = "https://arbetsformedlingen.se"
)
}
\arguments{
\item{url_webbsida}{URL till sidan med länkar.}

\item{sokord}{Sökord som identifierar rätt fil (se
\code{\link[=webbsida_extrahera_url_med_sokord]{webbsida_extrahera_url_med_sokord()}}).}

\item{bas_url}{Bas-URL för relativa länkar.}
}
\value{
En teckenvektor med sökväg(ar) till de nedladdade temporära filerna.
}
\description{
Letar bland länkarna på \code{url_webbsida} efter den (eller de) som matchar alla
\code{sokord} och laddar ner den till en temporär fil.
}
