% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{forandring_till_text}
\alias{forandring_till_text}
\title{Beskriv en förändring mellan två värden i text}
\usage{
forandring_till_text(varde1, varde2)
}
\arguments{
\item{varde1}{Startvärde.}

\item{varde2}{Slutvärde.}
}
\value{
En sträng som beskriver förändringen.
}
\description{
\code{forandring_till_text(100, 150)} ger \code{"ökat väsentligt"}.
}
