% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upprepa.R
\name{funktion_upprepa_forsok_tills_retur_TRUE}
\alias{funktion_upprepa_forsok_tills_retur_TRUE}
\title{Kör en funktion om och om igen tills den returnerar TRUE}
\usage{
funktion_upprepa_forsok_tills_retur_TRUE(
  funktion,
  max_forsok = 15,
  vanta_minuter = 2
)
}
\arguments{
\item{funktion}{En funktion utan argument som returnerar \code{TRUE}/\code{FALSE}.}

\item{max_forsok}{Max antal försök.}

\item{vanta_minuter}{Minuter att vänta mellan försöken.}
}
\value{
\code{TRUE} om något försök lyckades, annars \code{FALSE}.
}
\description{
Praktiskt för att vänta in en server eller ett jobb som blir klart.
}
