% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filer.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Negation av \code{\%in\%}}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{Värden som testas.}

\item{table}{Värden att testa mot.}
}
\value{
Logisk vektor.
}
\description{
\code{x \%notin\% tabell} är \code{!(x \%in\% tabell)}.
}
