% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skolverket.R
\name{gymnprg_inr_koder_hamta_api_skolverket}
\alias{gymnprg_inr_koder_hamta_api_skolverket}
\title{Hämta gymnasieprogrammens inriktningskoder från Skolverkets API}
\usage{
gymnprg_inr_koder_hamta_api_skolverket(
  url = "https://api.skolverket.se/planned-educations/v3/support/programs"
)
}
\arguments{
\item{url}{API-endpoint för program.}
}
\value{
En \code{tibble} med \code{Kod}, \code{Namn} och \code{skolform}, eller \code{NULL} vid fel.
}
\description{
Hämta gymnasieprogrammens inriktningskoder från Skolverkets API
}
