% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fk.R
\name{hamta_fk_json_dataset_med_url}
\alias{hamta_fk_json_dataset_med_url}
\title{Hämta ett dataset i JSON-format från Försäkringskassan via URL}
\usage{
hamta_fk_json_dataset_med_url(url_fk)
}
\arguments{
\item{url_fk}{URL till JSON-datasetet.}
}
\value{
En \code{tibble} med klartextkolumner och tabellnamn.
}
\description{
URL:er hittas på \url{https://www.dataportal.se} (filtrera på Försäkringskassan).
För Excel-filer, använd \code{\link[=hamta_excel_dataset_med_url]{hamta_excel_dataset_med_url()}} i stället.
}
