% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada.R
\name{hamta_kolada_giltiga_ar}
\alias{hamta_kolada_giltiga_ar}
\title{Giltiga år för ett Kolada-KPI}
\usage{
hamta_kolada_giltiga_ar(kpi_id, vald_region = "2080")
}
\arguments{
\item{kpi_id}{KPI-id, t.ex. \code{"N00945"}.}

\item{vald_region}{Kommunkod (nollutfylls till 4 tecken). Standard Falun.}
}
\value{
En teckenvektor med år.
}
\description{
Giltiga år för ett Kolada-KPI
}
