% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{hamta_kommunkoder}
\alias{hamta_kommunkoder}
\title{Kommunkoder med namn för ett län}
\usage{
hamta_kommunkoder(lanskod = "20", kolada = FALSE)
}
\arguments{
\item{lanskod}{Länskod, t.ex. \code{"20"}.}

\item{kolada}{Om \code{TRUE} nollutfylls koderna till 4 tecken.}
}
\value{
En \code{tibble} med \code{regionkod} och \code{region}.
}
\description{
Kommunkoder med namn för ett län
}
