% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{hamtaAllaLan}
\alias{hamtaAllaLan}
\title{Alla länskoder}
\usage{
hamtaAllaLan(tamedriket = TRUE)
}
\arguments{
\item{tamedriket}{Ta med rikskoden \code{"00"}.}
}
\value{
En teckenvektor med länskoder (och ev. \code{"00"}).
}
\description{
Alla länskoder
}
