% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{hamtakommuner}
\alias{hamtakommuner}
\title{Kommunkoder för ett eller flera län}
\usage{
hamtakommuner(
  lan = "20",
  tamedlan = TRUE,
  tamedriket = TRUE,
  allakommuner = FALSE
)
}
\arguments{
\item{lan}{Länskod(er), t.ex. \code{"20"}. Bara de två första tecknen används.}

\item{tamedlan}{Ta med länets egen kod i resultatet.}

\item{tamedriket}{Ta med rikskoden \code{"00"}.}

\item{allakommuner}{Om \code{TRUE} returneras alla kommuner i landet (\code{lan}
ignoreras för urvalet av kommuner).}
}
\value{
En teckenvektor med regionkoder.
}
\description{
Kommunkoder för ett eller flera län
}
