% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{hamtaregion_kod_namn}
\alias{hamtaregion_kod_namn}
\title{Slå upp region-namn för en eller flera regionkoder}
\usage{
hamtaregion_kod_namn(regionkod, kolada = FALSE)
}
\arguments{
\item{regionkod}{En eller flera regionkoder.}

\item{kolada}{Om \code{TRUE} nollutfylls koderna till 4 tecken (Kolada-format).}
}
\value{
En \code{tibble} med \code{regionkod} och \code{region}. Koder som inte hittas får
sig själva som \code{region}.
}
\description{
Slå upp region-namn för en eller flera regionkoder
}
