% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{hamtaregtab}
\alias{hamtaregtab}
\title{Regiontabell med kod och namn för riket, län och kommuner}
\usage{
hamtaregtab(uppdatera = FALSE)
}
\arguments{
\item{uppdatera}{Om \code{TRUE} hämtas tabellen om även om den finns i cachen.}
}
\value{
En \code{tibble} med kolumnerna \code{regionkod} och \code{region}.
}
\description{
Hämtar SCB:s standard-valuesets för riket, län och kommuner och sätter ihop
dem till en tabell. Resultatet cachas per R-session.
}
