% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verktyg.R
\name{hitta_funktioner_i_fil_ej_inuti_andra_funktioner}
\alias{hitta_funktioner_i_fil_ej_inuti_andra_funktioner}
\title{Lista bara toppnivåfunktioner i en R-fil}
\usage{
hitta_funktioner_i_fil_ej_inuti_andra_funktioner(filnamn)
}
\arguments{
\item{filnamn}{Sökväg eller URL till en R-fil.}
}
\value{
En teckenvektor med funktionsnamn.
}
\description{
Som \code{\link[=lista_funktioner_i_skript]{lista_funktioner_i_skript()}} men tar bara med funktioner som inte är
definierade inuti andra funktioner.
}
