% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_zip.R
\name{las_b64}
\alias{las_b64}
\title{Läs en base64-kodad fil och avkoda till text}
\usage{
las_b64(sokvag_filnamn)
}
\arguments{
\item{sokvag_filnamn}{Sökväg till filen.}
}
\value{
Det avkodade innehållet som en sträng.
}
\description{
Läs en base64-kodad fil och avkoda till text
}
