% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{list_komma_eller}
\alias{list_komma_eller}
\title{Sätt ihop en vektor till en text med kommatecken och "eller"}
\usage{
list_komma_eller(skickad_vektor)
}
\arguments{
\item{skickad_vektor}{En vektor som sätts ihop till en textsträng.}
}
\value{
En sträng med längd 1.
}
\description{
\code{list_komma_eller(c("a", "b", "c"))} ger \code{"a, b eller c"}.
}
