% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{list_komma_och}
\alias{list_komma_och}
\title{Sätt ihop en vektor till en text med kommatecken och "och"}
\usage{
list_komma_och(skickad_vektor)
}
\arguments{
\item{skickad_vektor}{En vektor som sätts ihop till en textsträng.}
}
\value{
En sträng med längd 1 (eller den oförändrade vektorn om den har
0-1 element).
}
\description{
\code{list_komma_och(c("a", "b", "c"))} ger \code{"a, b och c"}.
}
