% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{list_komma_samt}
\alias{list_komma_samt}
\title{Sätt ihop en vektor till en text med kommatecken och "samt"}
\usage{
list_komma_samt(skickad_vektor)
}
\arguments{
\item{skickad_vektor}{En vektor som sätts ihop till en textsträng.}
}
\value{
En sträng med längd 1.
}
\description{
\code{list_komma_samt(c("a", "b", "c"))} ger \code{"a, b samt c"}.
}
