% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verktyg.R
\name{lista_funktioner_i_skript}
\alias{lista_funktioner_i_skript}
\title{Lista funktionsnamn i en eller flera R-filer}
\usage{
lista_funktioner_i_skript(filsokvag_eller_github_url)
}
\arguments{
\item{filsokvag_eller_github_url}{En eller flera sökvägar eller URL:er till
R-filer.}
}
\value{
En teckenvektor med unika funktionsnamn.
}
\description{
Lista funktionsnamn i en eller flera R-filer
}
