% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappar.R
\name{mapp_hamtadata}
\alias{mapp_hamtadata}
\title{Sökväg till mappen för hämtad data}
\usage{
mapp_hamtadata()
}
\value{
En sökväg (sträng). Sätt eget värde med
\code{options(rdverktyg.mapp_hamtadata = "...")}.
}
\description{
Sökväg till mappen för hämtad data
}
