% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappar.R
\name{mapp_inlasdata}
\alias{mapp_inlasdata}
\title{Sökväg till mappen med inläsningsdata från mikrodatabasen}
\usage{
mapp_inlasdata()
}
\value{
En sökväg (sträng). Sätt eget värde med
\code{options(rdverktyg.mapp_inlasdata = "...")}.
}
\description{
Sökväg till mappen med inläsningsdata från mikrodatabasen
}
\details{
Sätt eget värde för den pågående sessionen med
\code{options(rdverktyg.mapp_inlasdata = "...")}. Vill du att värdet ska
gälla permanent, mellan R-sessioner, lägg samma \code{options()}-rad i din
\code{.Rprofile} - öppna den med \code{usethis::edit_r_profile()}, spara raden,
spara filen och starta om R.
}
