% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappar.R
\name{mapp_leveranser}
\alias{mapp_leveranser}
\title{Sökväg till leveransmappen}
\usage{
mapp_leveranser()
}
\value{
En sökväg (sträng). Sätt eget värde med
\code{options(rdverktyg.mapp_leveranser = "...")}.
}
\description{
Sökväg till leveransmappen
}
\details{
Sätt eget värde för den pågående sessionen med
\code{options(rdverktyg.mapp_leveranser = "...")}. Vill du att värdet ska
gälla permanent, mellan R-sessioner, lägg samma \code{options()}-rad i din
\code{.Rprofile} - öppna den med \code{usethis::edit_r_profile()}, spara raden,
spara filen och starta om R.
}
