% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappar.R
\name{mapp_temp}
\alias{mapp_temp}
\title{Sökväg till temp-mappen}
\usage{
mapp_temp()
}
\value{
En sökväg (sträng). Sätt eget värde med
\code{options(rdverktyg.mapp_temp = "...")}.
}
\description{
Sökväg till temp-mappen
}
