% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verktyg.R
\name{nummer_till_text}
\alias{nummer_till_text}
\title{Skriv ut tal 1-20 som svenska ord}
\usage{
nummer_till_text(x)
}
\arguments{
\item{x}{Heltal 1-20 (vektor). Övriga tal tas bort med en varning.}
}
\value{
En teckenvektor.
}
\description{
Skriv ut tal 1-20 som svenska ord
}
