% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{procent_till_text}
\alias{procent_till_text}
\title{Omvandla ett procenttal till text}
\usage{
procent_till_text(procent)
}
\arguments{
\item{procent}{Ett procenttal (0-100).}
}
\value{
En sträng.
}
\description{
\code{procent_till_text(50)} ger \code{"varannan"}, \code{procent_till_text(25)} ger
\code{"var fjärde"} och så vidare. Värden utan en bra textform returneras som
\code{"<x>\%"}.
}
