% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdverktyg-package.R
\docType{package}
\name{rdverktyg-package}
\alias{rdverktyg}
\alias{rdverktyg-package}
\title{rdverktyg: Allmänna Verktyg för Region Dalarnas Analysarbete}
\description{
Samling av allmänna hjälpfunktioner för text, filer, SCB-hämtning och andra återkommande behov i Region Dalarnas analysarbete. Utbrutet ur funktionsskripten i Region-Dalarna/funktioner.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Region-Dalarna/rdpaket}
  \item Report bugs at \url{https://github.com/Region-Dalarna/rdpaket/issues}
}

}
\author{
\strong{Maintainer}: Peter Möller \email{larspetermoller@gmail.com}

Authors:
\itemize{
  \item Peter Möller \email{larspetermoller@gmail.com}
}

Other contributors:
\itemize{
  \item Region Dalarna [copyright holder]
}

}
\keyword{internal}
