% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verktyg.R
\name{skapa_intervaller}
\alias{skapa_intervaller}
\title{Dela upp värden i ett antal intervall}
\usage{
skapa_intervaller(skickad_kolumn, antal_intervaller = 5)
}
\arguments{
\item{skickad_kolumn}{Numerisk vektor.}

\item{antal_intervaller}{Antal intervall (standard 5).}
}
\value{
En numerisk vektor med intervallgränser.
}
\description{
Delar spannet i \code{skickad_kolumn} i \code{antal_intervaller} steg avrundade till
"fina" tal. Praktiskt för legender och storleksklasser.
}
