% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{skapa_kortnamn_lan}
\alias{skapa_kortnamn_lan}
\title{Kortnamn för län}
\usage{
skapa_kortnamn_lan(lansnamn, byt_ut_riket_mot_sverige = FALSE)
}
\arguments{
\item{lansnamn}{Ett eller flera länsnamn.}

\item{byt_ut_riket_mot_sverige}{Om \code{TRUE} blir \code{"Riket"} -> \code{"Sverige"}.}
}
\value{
En teckenvektor med kortnamn.
}
\description{
\code{skapa_kortnamn_lan("Dalarnas län")} ger \code{"Dalarna"}. Tar bort avslutande
\code{" län"} och ett ev. foge-s.
}
