% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filer.R
\name{skapa_mapp_om_den_inte_finns}
\alias{skapa_mapp_om_den_inte_finns}
\title{Skapa en mapp om den inte redan finns}
\usage{
skapa_mapp_om_den_inte_finns(sokvag)
}
\arguments{
\item{sokvag}{Sökväg till mappen. Skapas rekursivt.}
}
\value{
\code{sokvag} (osynligt).
}
\description{
Skapa en mapp om den inte redan finns
}
