% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verktyg.R
\name{slash_lagg_till}
\alias{slash_lagg_till}
\title{Lägg till avslutande snedstreck om det saknas}
\usage{
slash_lagg_till(x)
}
\arguments{
\item{x}{En eller flera sökvägar/strängar.}
}
\value{
\code{x} med avslutande \code{/}.
}
\description{
Lägg till avslutande snedstreck om det saknas
}
