% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verktyg.R
\name{sokvag_for_skript_hitta}
\alias{sokvag_for_skript_hitta}
\title{Sökväg till det skript som körs}
\usage{
sokvag_for_skript_hitta()
}
\value{
En sökväg som slutar med \code{/}.
}
\description{
Försöker i tur och ordning: \code{this.path}, \code{source()}-kontext,
RStudio-dokumentet, och till sist arbetskatalogen.
}
