% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{source_funktioner}
\alias{source_funktioner}
\title{Source:a bara en eller flera namngivna funktioner från en fil}
\usage{
source_funktioner(skriptfil, funktioner)
}
\arguments{
\item{skriptfil}{Sökväg eller URL till en R-fil.}

\item{funktioner}{Teckenvektor med funktionsnamn att plocka ut.}
}
\value{
Inget - de hittade funktionerna läggs i den globala miljön.
}
\description{
Source:a bara en eller flera namngivna funktioner från en fil
}
