% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_zip.R
\name{spara_som_csv_i_zip}
\alias{spara_som_csv_i_zip}
\title{Spara en eller flera data.frames som csv-filer i en zip}
\usage{
spara_som_csv_i_zip(
  df_list,
  output_mapp = NULL,
  zipfilnamn = NA,
  pre_namn_csv_fil_utan_namn = "df_",
  meddelande = TRUE
)
}
\arguments{
\item{df_list}{En \code{data.frame} eller en (namngiven) lista av \code{data.frame}s.}

\item{output_mapp}{Målmapp. \code{NULL} = \code{\link[=utskriftsmapp]{utskriftsmapp()}}.}

\item{zipfilnamn}{Namn på zip-filen. \code{NA} = härled.}

\item{pre_namn_csv_fil_utan_namn}{Prefix för csv-filer vars listelement saknar namn.}

\item{meddelande}{Om \code{TRUE} skrivs en bekräftelse ut.}
}
\value{
Inget - skriver zip-filen.
}
\description{
Spara en eller flera data.frames som csv-filer i en zip
}
