% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verktyg.R
\name{suppress_specific_warning}
\alias{suppress_specific_warning}
\title{Dämpa varningar som matchar en text}
\usage{
suppress_specific_warning(expr, warn_text = "NAs introduced by coercion")
}
\arguments{
\item{expr}{Uttrycket som ska köras.}

\item{warn_text}{Delsträng - varningar vars meddelande innehåller den dämpas.}
}
\value{
Värdet av \code{expr}.
}
\description{
Dämpa varningar som matchar en text
}
