% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{tatortskoder_bearbeta}
\alias{tatortskoder_bearbeta}
\title{Utöka läns-/kommunkoder till tätortskoder}
\usage{
tatortskoder_bearbeta(
  tabell_id,
  koder,
  region_var = "Region",
  base_url = "https://statistikdatabasen.scb.se/api/v2/tables/"
)
}
\arguments{
\item{tabell_id}{PxWeb-tabell-id.}

\item{koder}{Läns- (2 tecken), kommun- (4 tecken) eller tätortskoder
(>4 tecken). \code{"*"} returneras oförändrat.}

\item{region_var}{Regionvariabelns namn i tabellen.}

\item{base_url}{Bas-URL till PxWeb API v2.}
}
\value{
En teckenvektor med tätortskoder.
}
\description{
Skicka läns-, kommun- eller tätortskoder och få tillbaka alla giltiga
tätortskoder inom dem (från den angivna tabellen).
}
