% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webb.R
\name{url_finns_webbsida}
\alias{url_finns_webbsida}
\title{Svarar en webbsida med HTTP 200?}
\usage{
url_finns_webbsida(skickad_url)
}
\arguments{
\item{skickad_url}{URL att testa.}
}
\value{
\code{TRUE} om ett HEAD-anrop ger status 200.
}
\description{
Svarar en webbsida med HTTP 200?
}
