% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_zip.R
\name{csv_fran_zipfiler_inlasning}
\alias{csv_fran_zipfiler_inlasning}
\title{Läs csv-filer direkt ur en eller flera zip-filer}
\usage{
csv_fran_zipfiler_inlasning(zip_sokvagar, kalla_som_kolumn = FALSE)
}
\arguments{
\item{zip_sokvagar}{Sökväg(ar) till zip-filer som innehåller csv-filer.}

\item{kalla_som_kolumn}{Om \code{TRUE} läggs kolumner \code{zip_fil} och \code{csv_fil} till
som visar varifrån varje rad kommer.}
}
\value{
En \code{tibble} med alla csv-filers innehåll ihopbundet.
}
\description{
Läs csv-filer direkt ur en eller flera zip-filer
}
