% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappar.R
\name{mapp_hamtadata}
\alias{mapp_hamtadata}
\title{Sökväg till mappen för hämtad data}
\usage{
mapp_hamtadata()
}
\value{
En sökväg (sträng). Sätt eget värde med
\code{options(rdverktyg.mapp_hamtadata = "...")}.
}
\description{
Sökväg till mappen för hämtad data
}
\details{
Sätt eget värde för den pågående sessionen med
\code{options(rdverktyg.mapp_hamtadata = "...")}. Vill du att värdet ska
gälla permanent, mellan R-sessioner, lägg samma \code{options()}-rad i din
\code{.Rprofile} - öppna den med \code{usethis::edit_r_profile()}, spara raden,
spara filen och starta om R.
}
