% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappar.R
\name{mapp_inlasdata}
\alias{mapp_inlasdata}
\title{Sökväg till mappen med inläsningsdata från mikrodatabasen}
\usage{
mapp_inlasdata()
}
\value{
En sökväg (sträng). Sätt eget värde med
\code{options(rdverktyg.mapp_inlasdata = "...")}.
}
\description{
Sökväg till mappen med inläsningsdata från mikrodatabasen
}
