% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappar.R
\name{mapp_leveranser}
\alias{mapp_leveranser}
\title{Sökväg till leveransmappen}
\usage{
mapp_leveranser()
}
\value{
En sökväg (sträng). Sätt eget värde med
\code{options(rdverktyg.mapp_leveranser = "...")}.
}
\description{
Sökväg till leveransmappen
}
