% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappar.R
\name{utskriftsmapp}
\alias{utskriftsmapp}
\title{Sökväg till utskriftsmappen}
\usage{
utskriftsmapp()
}
\value{
En sökväg (sträng). Sätt eget värde med
\code{options(rdverktyg.utskriftsmapp = "...")}.
}
\description{
Sökväg till utskriftsmappen
}
